Move the event bus classes into their own package so that they can be reused.
