+++ grader/src/main/java/edu/pdx/cs410J/grader/mvp/EventBusThatPublishesUnhandledExceptionEvents.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.grader.poa;
+package edu.pdx.cs410J.grader.mvp;
 
 import com.google.common.eventbus.EventBus;
 import com.google.common.eventbus.Subscribe;
