+++ grader/src/main/java/edu/pdx/cs410J/grader/poa/ui/POAGraderUIModule.java
@@ -4,6 +4,9 @@ import com.google.common.eventbus.EventBus;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
+import edu.pdx.cs410J.grader.mvp.EventBusThatPublishesUnhandledExceptionEvents;
+import edu.pdx.cs410J.grader.mvp.UnhandledExceptionPresenter;
+import edu.pdx.cs410J.grader.mvp.UnhandledExceptionView;
 import edu.pdx.cs410J.grader.poa.*;
 
 public class POAGraderUIModule extends AbstractModule {
